feat(ehr): moving shared code
Ref: #1040
