+++ packages/api/src/domain/cx-mapping.ts
@@ -1,18 +1,20 @@
 import { BaseDomain } from "@metriport/core/domain/base-domain";
-import { ehrSources } from "@metriport/shared/src/interface/external/shared/ehr/source";
+import { ehrSources } from "@metriport/shared/src/interface/external/ehr/source";
 import { z } from "zod";
-import { ehrCxMappingSecondaryMappingsSchemaMap } from "../external/ehr/shared";
+import {
+  EhrCxMappingSecondaryMappings,
+  ehrCxMappingSecondaryMappingsSchemaMap,
+} from "../external/ehr/shared";
 
 const cxMappingSource = [...ehrSources] as const;
 export type CxMappingSource = (typeof cxMappingSource)[number];
 export function isCxMappingSource(source: string): source is CxMappingSource {
   return cxMappingSource.includes(source as CxMappingSource);
 }
-
-export type CxMappingSecondaryMappings = {
+export type CxMappingSecondaryMappings = EhrCxMappingSecondaryMappings | null;
+export const secondaryMappingsSchemaMap: {
   [key in CxMappingSource]: z.Schema | undefined;
-};
-export const secondaryMappingsSchemaMap: CxMappingSecondaryMappings = {
+} = {
   ...ehrCxMappingSecondaryMappingsSchemaMap,
 };
