+++ packages/api/src/domain/patient-mapping.ts
@@ -1,5 +1,5 @@
 import { BaseDomain } from "@metriport/core/domain/base-domain";
-import { ehrSources } from "@metriport/shared/src/interface/external/shared/ehr/source";
+import { ehrSources } from "@metriport/shared/src/interface/external/ehr/source";
 
 export type PatientSourceIdentifierMap = {
   [key in string]: string[];
