+++ packages/api/src/external/ehr/athenahealth/command/process-patients-from-appointments.ts
@@ -4,8 +4,8 @@ import { executeAsynchronously } from "@metriport/core/util/concurrency";
 import { out } from "@metriport/core/util/log";
 import { capture } from "@metriport/core/util/notifications";
 import { MetriportError, errorToString } from "@metriport/shared";
-import { BookedAppointment } from "@metriport/shared/src/interface/external/athenahealth/index";
-import { EhrSources } from "@metriport/shared/src/interface/external/shared/ehr/source";
+import { BookedAppointment } from "@metriport/shared/src/interface/external/ehr/athenahealth";
+import { EhrSources } from "@metriport/shared/src/interface/external/ehr/source";
 import dayjs from "dayjs";
 import duration from "dayjs/plugin/duration";
 import { uniqBy } from "lodash";
