+++ packages/api/src/external/ehr/athenahealth/command/sync-patient.ts
@@ -1,6 +1,6 @@
 import AthenaHealthApi from "@metriport/core/external/ehr/athenahealth";
 import { isAthenaCustomFieldsEnabledForCx } from "@metriport/core/external/aws/app-config";
-import { EhrSources } from "@metriport/shared/src/interface/external/shared/ehr/source";
+import { EhrSources } from "@metriport/shared/src/interface/external/ehr/source";
 import { processAsyncError } from "@metriport/core/util/error/shared";
 import { BadRequestError } from "@metriport/shared";
 import { findOrCreatePatientMapping, getPatientMapping } from "../../../../command/mapping/patient";
