+++ packages/api/src/external/ehr/athenahealth/command/write-condition-to-chart.ts
@@ -1,5 +1,5 @@
 import { Condition } from "@medplum/fhirtypes";
-import { CreatedProblemSuccess } from "@metriport/shared/interface/external/athenahealth/problem";
+import { CreatedProblemSuccess } from "@metriport/shared/interface/external/ehr/athenahealth/problem";
 import { createAthenaClient } from "../shared";
 
 export async function writeConditionToChart({
