+++ packages/api/src/external/ehr/athenahealth/command/write-vitals-to-chart.ts
@@ -1,5 +1,5 @@
 import { GroupedVitals } from "@metriport/core/external/ehr/athenahealth";
-import { CreatedVitalsSuccess } from "@metriport/shared/interface/external/athenahealth/vitals";
+import { CreatedVitalsSuccess } from "@metriport/shared/interface/external/ehr/athenahealth/vitals";
 import { createAthenaClient } from "../shared";
 
 export async function writeVitalsToChart({
