+++ packages/api/src/external/ehr/canvas/command/sync-patient.ts
@@ -1,5 +1,5 @@
 import CanvasApi from "@metriport/core/external/canvas/index";
-import { EhrSources } from "@metriport/shared/src/interface/external/shared/ehr/source";
+import { EhrSources } from "@metriport/shared/src/interface/external/ehr/source";
 import { processAsyncError } from "@metriport/core/util/error/shared";
 import { findOrCreatePatientMapping, getPatientMapping } from "../../../../command/mapping/patient";
 import { queryDocumentsAcrossHIEs } from "../../../../command/medical/document/document-query";
