+++ packages/api/src/external/ehr/shared.ts
@@ -3,13 +3,16 @@ import CanvasApi, { CanvasEnv } from "@metriport/core/external/ehr/canvas/index"
 import ElationApi, { ElationEnv } from "@metriport/core/external/ehr/elation";
 import { JwtTokenInfo, MetriportError } from "@metriport/shared";
 import { buildDayjs } from "@metriport/shared/common/date";
-import { athenaSecondaryMappingsSchema } from "@metriport/shared/interface/external/athenahealth/cx-mapping";
+import {
+  athenaSecondaryMappingsSchema,
+  AthenaSecondaryMappings,
+} from "@metriport/shared/interface/external/ehr/athenahealth/cx-mapping";
 import {
   AthenaClientJwtTokenData,
   athenaClientSource,
   AthenaDashJwtTokenData,
   athenaDashSource,
-} from "@metriport/shared/src/interface/external/athenahealth/jwt-token";
+} from "@metriport/shared/src/interface/external/ehr/athenahealth/jwt-token";
 import {
   CanvasClientJwtTokenData,
   canvasClientSource,
@@ -17,14 +20,14 @@ import {
   canvasDashSource,
   CanvasWebhookJwtTokenData,
   canvasWebhookSource,
-} from "@metriport/shared/src/interface/external/canvas/jwt-token";
+} from "@metriport/shared/src/interface/external/ehr/canvas/jwt-token";
 import {
   ElationClientJwtTokenData,
   elationClientSource,
   ElationWebhookJwtTokenData,
   elationWebhookSource,
-} from "@metriport/shared/src/interface/external/elation/jwt-token";
-import { EhrSource, EhrSources } from "@metriport/shared/src/interface/external/shared/ehr/source";
+} from "@metriport/shared/src/interface/external/ehr/elation/jwt-token";
+import { EhrSource, EhrSources } from "@metriport/shared/src/interface/external/ehr/source";
 import dayjs from "dayjs";
 import { Duration } from "dayjs/plugin/duration";
 import { z } from "zod";
@@ -81,6 +84,7 @@ export function isEhrWebhookJwtTokenSource(source: string): source is EhrWebhook
 
 export type EhrWebhookJwtTokenData = CanvasWebhookJwtTokenData | ElationWebhookJwtTokenData;
 
+export type EhrCxMappingSecondaryMappings = AthenaSecondaryMappings;
 export const ehrCxMappingSecondaryMappingsSchemaMap: {
   [key in EhrSource]: z.Schema | undefined;
 } = {
