+++ packages/shared/src/interface/external/ehr/elation/appointment.ts
